(minor) add ImageButton, add LocalesState, add dynamic i18n setup and changing

+++ class.sh
@@ -11,10 +11,17 @@ then
         mkdir "include/$1"
     fi
 
-    touch "src/${1}/${2}.cxx"
-    touch "include/${1}/${2}.hxx"
-    echo -e "#include \"stdafx.hxx\"\n#include \"$1/$2.hxx\"\n" > "src/${1}/${2}.cxx"
-    echo -e "#pragma once" > "include/${1}/${2}.hxx"
+    if [ ! -f "src/${1}/${2}.cxx" ]
+    then
+        touch "src/${1}/${2}.cxx"
+        echo -e "#include \"stdafx.hxx\"\n#include \"$1/$2.hxx\"\n" > "src/${1}/${2}.cxx"
+    fi
+
+    if [ ! -f "include/${1}/${2}.hxx" ]
+    then
+        touch "include/${1}/${2}.hxx"
+        echo -e "#pragma once" > "include/${1}/${2}.hxx"
+    fi
 
     ./updatesrcdeps.sh
 else
